+++ lib/dotenv/cli.rb
@@ -4,77 +4,54 @@ require "dotenv/template"
 require "optparse"
 
 module Dotenv
-  # The CLI is a class responsible of handling all the command line interface
-  # logic.
-  class CLI
-    attr_reader :argv, :filenames
+  # The command line interface
+  class CLI < OptionParser
+    attr_reader :argv, :filenames, :overload
 
     def initialize(argv = [])
       @argv = argv.dup
       @filenames = []
-    end
+      @overload = false
 
-    def run
-      parse_argv!(@argv)
+      super "Usage: dotenv [options]"
+      separator ""
 
-      begin
-        Dotenv.load!(*@filenames)
-      rescue Errno::ENOENT => e
-        abort e.message
-      else
-        exec(*@argv) unless @argv.empty?
-      end
-    end
-
-    private
-
-    def parse_argv!(argv)
-      parser = create_option_parser
-      add_options(parser)
-      parser.order!(argv)
-
-      @filenames
-    end
-
-    def add_options(parser)
-      add_files_option(parser)
-      add_help_option(parser)
-      add_version_option(parser)
-      add_template_option(parser)
-    end
-
-    def add_files_option(parser)
-      parser.on("-f FILES", Array, "List of env files to parse") do |list|
+      on("-f FILES", Array, "List of env files to parse") do |list|
         @filenames = list
       end
-    end
 
-    def add_help_option(parser)
-      parser.on("-h", "--help", "Display help") do
-        puts parser
+      on("-o", "--overload", "override existing ENV variables") do
+        @overload = true
+      end
+
+      on("-h", "--help", "Display help") do
+        puts self
         exit
       end
-    end
 
-    def add_version_option(parser)
-      parser.on("-v", "--version", "Show version") do
+      on("-v", "--version", "Show version") do
         puts "dotenv #{Dotenv::VERSION}"
         exit
       end
-    end
 
-    def add_template_option(parser)
-      parser.on("-t", "--template=FILE", "Create a template env file") do |file|
+      on("-t", "--template=FILE", "Create a template env file") do |file|
         template = Dotenv::EnvTemplate.new(file)
         template.create_template
       end
+
+      order!(@argv)
     end
 
-    def create_option_parser
-      OptionParser.new do |parser|
-        parser.banner = "Usage: dotenv [options]"
-        parser.separator ""
+    def run
+      if @overload
+        Dotenv.overload!(*@filenames)
+      else
+        Dotenv.load!(*@filenames)
       end
+    rescue Errno::ENOENT => e
+      abort e.message
+    else
+      exec(*@argv) unless @argv.empty?
     end
   end
 end